Adapted IntegrationTestBase to new model
