+++ LoRaEngine/test/LoRaWan.IntegrationTest/IntegrationTestBaseCi.cs
@@ -1,0 +1,51 @@
+﻿using LoRaWan.Test.Shared;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Text;
+using Xunit;
+
+namespace LoRaWan.IntegrationTest
+{
+    public class IntegrationTestBaseCi : IntegrationTestBase, IClassFixture<IntegrationTestFixtureCi>, IDisposable
+    {        
+        protected IntegrationTestFixtureCi TestFixtureCi { get { return (IntegrationTestFixtureCi)this.TestFixture; } }
+      
+        private LoRaArduinoSerial arduinoDevice;
+        protected LoRaArduinoSerial ArduinoDevice { get { return this.arduinoDevice; } }
+
+        public IntegrationTestBaseCi(IntegrationTestFixtureCi testFixture)
+            : base(testFixture)
+        {
+            this.arduinoDevice = testFixture.ArduinoDevice;
+        }
+
+        private bool disposedValue = false; // To detect redundant calls
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (!disposedValue)
+            {
+                if (disposing)
+                {
+                    // Before starting a new test, wait 5 seconds to ensure serial port is not receiving dirty data
+                    if (this.arduinoDevice != null)
+                        this.arduinoDevice.WaitForIdleAsync(TimeSpan.FromSeconds(5)).GetAwaiter().GetResult();
+                }
+
+                disposedValue = true;
+            }
+        }
+
+
+        // This code added to correctly implement the disposable pattern.
+        public void Dispose()
+        {
+            // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
+            Dispose(true);
+            // TODO: uncomment the following line if the finalizer is overridden above.
+            GC.SuppressFinalize(this);
+        }
+    }
+}
