--- LoRaEngine/test/LoRaWan.IntegrationTest/IntegrationTestFixture.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace LoRaWan.IntegrationTest
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Text;
-
-    public class IntegrationTestFixtureCi : IntegrationTestFixtureBase
-    {
-    }
-}
