+++ LoRaEngine/test/LoRaWan.IntegrationTest/SensorDecodingTest.cs
@@ -9,7 +9,7 @@ namespace LoRaWan.IntegrationTest
     // Tests sensor decoding test (http, reflection)
     [Collection(Constants.TestCollectionName)] // run in serial
     [Trait("Category", "SkipWhenLiveUnitTesting")]
-    public class SensorDecodingTest: IntegrationTestBase
+    public class SensorDecodingTest: IntegrationTestBaseCi
     {
         public SensorDecodingTest(IntegrationTestFixtureCi testFixture) : base(testFixture)
         {
@@ -21,14 +21,14 @@ namespace LoRaWan.IntegrationTest
         [Fact]    
         public async Task SensorDecoder_HttpBased_ValueSensorDecoder_DecodesPayload()
         {
-            var device = this.TestFixture.Device11_OTAA;
+            var device = this.TestFixtureCi.Device11_OTAA;
             LogTestStart(device);    
 
             await this.ArduinoDevice.setDeviceModeAsync(LoRaArduinoSerial._device_mode_t.LWOTAA);
             await this.ArduinoDevice.setIdAsync(device.DevAddr, device.DeviceID, device.AppEUI);
             await this.ArduinoDevice.setKeyAsync(device.NwkSKey, device.AppSKey, device.AppKey);
 
-            await this.ArduinoDevice.SetupLora(this.TestFixture.Configuration.LoraRegion);
+            await this.ArduinoDevice.SetupLora(this.TestFixtureCi.Configuration.LoraRegion);
 
             var joinSucceeded = await this.ArduinoDevice.setOTAAJoinAsyncWithRetry(LoRaArduinoSerial._otaa_join_cmd_t.JOIN, 20000, 5);
             Assert.True(joinSucceeded, "Join failed");
@@ -44,9 +44,9 @@ namespace LoRaWan.IntegrationTest
             await AssertUtils.ContainsWithRetriesAsync("+CMSG: ACK Received", this.ArduinoDevice.SerialLogs);
 
             // Find "0000000000000011: message '{"value":1234}' sent to hub" in network server logs
-            await this.TestFixture.AssertNetworkServerModuleLogStartsWithAsync($"{device.DeviceID}: message '{{\"value\":\"1234\"}}' sent to hub");
+            await this.TestFixtureCi.AssertNetworkServerModuleLogStartsWithAsync($"{device.DeviceID}: message '{{\"value\":\"1234\"}}' sent to hub");
 
-            this.TestFixture.ClearLogs();
+            this.TestFixtureCi.ClearLogs();
         }
 
         // Ensures that reflect based sensor decoder decodes payload
@@ -54,14 +54,14 @@ namespace LoRaWan.IntegrationTest
         [Fact]
         public async Task SensorDecoder_ReflectionBased_ValueSensorDecoder_DecodesPayload()
         {
-            var device = this.TestFixture.Device12_OTAA;
+            var device = this.TestFixtureCi.Device12_OTAA;
             LogTestStart(device);      
 
             await this.ArduinoDevice.setDeviceModeAsync(LoRaArduinoSerial._device_mode_t.LWOTAA);
             await this.ArduinoDevice.setIdAsync(device.DevAddr, device.DeviceID, device.AppEUI);
             await this.ArduinoDevice.setKeyAsync(device.NwkSKey, device.AppSKey, device.AppKey);
 
-            await this.ArduinoDevice.SetupLora(this.TestFixture.Configuration.LoraRegion);
+            await this.ArduinoDevice.SetupLora(this.TestFixtureCi.Configuration.LoraRegion);
 
             var joinSucceeded = await this.ArduinoDevice.setOTAAJoinAsyncWithRetry(LoRaArduinoSerial._otaa_join_cmd_t.JOIN, 20000, 5);
             Assert.True(joinSucceeded, "Join failed");
@@ -78,9 +78,9 @@ namespace LoRaWan.IntegrationTest
             await AssertUtils.ContainsWithRetriesAsync("+CMSG: ACK Received", this.ArduinoDevice.SerialLogs);
 
             // Find "0000000000000011: message '{"value":1234}' sent to hub" in network server logs
-            await this.TestFixture.AssertNetworkServerModuleLogStartsWithAsync($"{device.DeviceID}: message '{{\"value\":4321}}' sent to hub");
+            await this.TestFixtureCi.AssertNetworkServerModuleLogStartsWithAsync($"{device.DeviceID}: message '{{\"value\":4321}}' sent to hub");
 
-            this.TestFixture.ClearLogs();
+            this.TestFixtureCi.ClearLogs();
         }
     }
